Move prove route to a server-routes crate
