+++ server-routes/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "server-routes"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+polylang-prover = { path = "../prover" }
+abi = { path = "../abi" }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+error = { path = "../error" }
+base64 = "0.21.4"
+tokio = { version = "1" }
