+++ server/Cargo.toml
@@ -7,10 +7,5 @@ edition = "2021"
 
 [dependencies]
 actix-web = "4.4.0"
-polylang-prover = { path = "../prover" }
-abi = { path = "../abi" }
-serde = { version = "1", features = ["derive"] }
 tokio = { version = "1.32.0", features = ["macros", "rt", "rt-multi-thread"] }
-serde_json = "1"
-error = { path = "../error" }
-base64 = "0.21.4"
+server-routes = { path = "../server-routes" }
